Example
Types of systems we will need to maintain the scalability of this project. Scalability is the key. Low power low resource over vast distances.

Be the Gravity of Data Processing. Weak acutely, but hits like a hammer at scale. The more uncorrupted unobserved data that we can process whose output is never actually seen by human eyes directly the better. That means 0 data corruption due to the destruction of a the wave form by direct observation. Humans never see the source of the qbit data packet. Only the output provided by any number of objects not including another qunatum device.